--- src/movekit/io/__init__.py
@@ -1,3 +1,0 @@
-from .excel import *
-from .csv import *
-name = 'io'
