--- src/movekit/preprocess/__init__.py
@@ -1,3 +1,0 @@
-from .preprocess import *
-from .interpolation import *
-name = 'preprocess'
